task 6 int -> double

+++ Lab1/src/bsuir/webtech/lab1/task6/Program.java
@@ -9,18 +9,18 @@ public class Program {
         System.out.print("N: ");
         int n = scan.nextInt();
 
-        int[] mas = new int[n];
+        double[] mas = new double[n];
         for (int i = 0; i < n; i++) {
             System.out.printf("A[%d]: ", i);
-            mas[i] = scan.nextInt();
+            mas[i] = scan.nextDouble();
         }
 
         printMatrix(getMatrix(mas));
     }
 
-    private static int[][] getMatrix(int[] mas) {
+    private static double[][] getMatrix(double[] mas) {
         int n = mas.length;
-        int[][] mat = new int[n][n];
+        double[][] mat = new double[n][n];
 
         for (int i = 0; i < n; i++) {
             for (int j = 0, k = i; k < n; j++, k++) {
@@ -34,10 +34,10 @@ public class Program {
         return mat;
     }
 
-    private static void printMatrix(int[][] mat) {
+    private static void printMatrix(double[][] mat) {
         for (int i = 0; i < mat.length; i++) {
             for (int j = 0; j < mat[i].length; j++) {
-                System.out.printf("%d ", mat[i][j]);
+                System.out.printf("%8.2f ", mat[i][j]);
             }
             System.out.println();
         }
